Remove numpy as dependency to setup.py
When importing the __version__ variable in the setup.py script an
import chain was triggered resulting in dependency of numpy leading to a
failed installtion of vi3o. Version is now read using exec on a specific
version file to avoid execution of the __init__.py in the main package.

Change-Id: Ic077320d89c63cfd7630c8b56c6cbad2894e94ed