Aggressive try to decode video
Files that didn't end with the correct postfix failed to be parsed by
vi3o. In some cases a file was parsed by vi3o if it had one postfix and
silently parsed by imageio if it had another postfix. Instead of
checking the postfix, try to decode and ask for forgivness if failing.