Repair AxisCam (imwrite fp interface)
Usage of vi3o.image.imwrite with open file descriptors was changed in
an earlier commit aiming to make vi3o compatible with pathlib. This
commit makes the vi3o.netcam.AxisCam class work again and also adds
test cases for this use.

(Note that pillow has been Pathlib compatible for a long time, but not
pathlib2 compatible, so the conversions are needed for python2)