Make window coordinate compensation exposed
When overloading on_mouse_press() and on_mouse_release() in an
DebugViewer derived class the x and y coordinates are in window
coordinate system, not in the image coordinate system. This is possible
to side step by using the already converted mouse_x and mouse_y
attributes, but these are class attributes, not instance attributes.
Also when using OpenGL methods in combination with vi3o the conversion
is usefull.

This commit exposes the conversion done in on_mouse_motion() to be
calleble by a subclass.

Change-Id: I16011e767ba6ae03de24d1cc5b246780ebb00840