Fix length property of mp4 video
It seems like the ffmpeg reader can't get the number of frames in an
.mp4 video. Using len() of the reader as we do now returns a large
(bogus?) number of frames. Asking for the the nframes meta data returns
inf. This commit deals with it in first hand by using ffprobe and in
second hand by trying to count the frames. This could take a very long
time and the stream could even be an infinitely long live stream, in
that case the len() function will hang forever.

It does however make the API work as one would assume it to work.

Change-Id: Ibc945ab3dfea1953b529a87f907e0bab899aee53