Add a ffprobe python wrapper
This commit adds a wrapper / reader for the ffprobe binary.

The code is a modification of an open source (MIT) project that is no
longer maintained.

The class assumes that the ffprobe binary exists in the path, which is
not very reliable. imageio-ffmpeg has similar problems, but does seem to
sometimes pack its own ffmpeg binary with the distribution. Thus one
should be careful when using this class.

Change-Id: Id14eb3a182ba79b5f2ed5daa7c06ff7375671fa1