Draw 1 channel images as gray in debugviewer
An image with shape (h, w, 1) should be rendered as a gray image. For
cases where the number of channels is not 1 or 3 it is more unclear what
would be the best action to perform, this implementation falls back to
gray which I think makes it easier to understand the actual structure of
the frame.